requirements.txt
python-dateutil==2.8.2
python-evtx==0.7.4
six==1.16.0
